% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFilt.R
\docType{package}
\name{CFilt}
\alias{CFilt}
\title{CFilt: A package about Collaborative Filtering by RC in R.}
\description{
The CFilt package provides one builder function CFbuilder and one class CF with methods that serve to change objects and recommend items or users.
}
\details{
Two main goals:
\itemize{
\item Structure the database so that changes can be made in a practical way through object-oriented programming.
\item Make recommendations through choices by the Collaborative Filtering methodology in a practical, fast and efficient manner.
}
}
\author{
Authors:
\itemize{
\item Jessica Quintanilha Kubrusly - jessicakubrusly@id.uff.br
\item Thiago Augusto Santos Lima - thiagoaugusto@id.uff.br
}
}
