% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cosine}
\alias{cosine}
\title{Similarity calculation Function}
\usage{
cosine(CF, type, i, j)
}
\arguments{
\item{CF}{A CF objec}

\item{type}{"user" or "item"}

\item{i}{"user" or "item" Id or index}

\item{j}{"user" or "item" Id or index#'}
}
\description{
Functions that returns the cosine similarity between two items or users.
}
\examples{
objectCF_r <- CFbuilder(Data = movies[1:500,], Datatype = "ratings", 
similarity = "cosine")
cosine(CF=objectCF_r,type = "user",i="1",j="2")
cosine(CF=objectCF_r,type = "item",i="Her",j="Frozen")
}
\author{
Jessica Kubrusly
}
