% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movie ratings by users}
\format{
A data frame with 7276 rows and 3 variables:
\describe{
\item{IdUsers}{Users identifier. Numbers 1 to 526.}
\item{IdItems}{Movies identifier. Movies names.}
\item{Ratings}{Movie ratings by users. Ratings from 1 to 5 (Likert scale).}
}
}
\source{
Giglio (2014)
}
\description{
A dataset containing 7276 ratings for 50 movies by 526 users.
}
\keyword{datasets}
