% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson.R
\name{pearson}
\alias{pearson}
\title{Similarity calculation Function}
\usage{
pearson(CF, type, i, j)
}
\arguments{
\item{CF}{A CF objec}

\item{type}{"user" or "item"}

\item{i}{"user" or "item" Id or index}

\item{j}{"user" or "item" Id or index#'}
}
\description{
Functions that returns the pearson similarity between two items or users.
}
\examples{
objectCF_r <- CFbuilder(Data = movies[1:500,], Datatype = "ratings", 
similarity = "pearson")
pearson(CF=objectCF_r,type = "user",i="2",j="3")
pearson(CF=objectCF_r,type = "item",i="Her",j="Frozen")
}
\author{
Jessica Kubrusly
}
