% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topkusers.R
\name{topkusers}
\alias{topkusers}
\title{Top-K User Recommendation for an Item}
\usage{
topkusers(CF, Id_i, k = 10, type = "user")
}
\arguments{
\item{CF}{An object of class \code{CF} created by \code{\link{CFbuilder}}.}

\item{Id_i}{A character string representing the item ID.}

\item{k}{A positive integer indicating the number of users to return.
Default is 10.}

\item{type}{A character string indicating the recommendation strategy:
\itemize{
\item \code{"user"}: user-based collaborative filtering
\item \code{"item"}: item-based collaborative filtering
}}
}
\value{
A character vector containing the IDs of the top-k users
for whom the item is recommended.
}
\description{
Returns the top-k users to whom a given item should be recommended,
based on a collaborative filtering model.
}
\details{
For \code{type = "user"}, recommendations are based on similarities
between users. For \code{type = "item"}, recommendations are based on
similarities between items.

Only users who have not yet consumed/rated the item are considered.
}
\examples{
data(movies, package = "CFilt")

CF1 <- CFbuilder(movies[1:200, ], Datatype = "rating")

# Recommend users for an item using user-based CF
topkusers(CF1, Id_i = "Frozen", k = 5, type = "user")

# Recommend users for an item using item-based CF
topkusers(CF1, Id_i = "Frozen", k = 5, type = "item")

CF2 <- CFbuilder(movies[1:200,-3])

# Recommend users for an item using user-based CF
topkusers(CF2, Id_i = "Frozen", k = 5, type = "user")

# Recommend users for an item using item-based CF
topkusers(CF2, Id_i = "Frozen", k = 3, type = "item")

}
\seealso{
\code{\link{CFbuilder}}, \code{\link{topkitems}}
}
