% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclosestitems.R
\name{kclosestitems}
\alias{kclosestitems}
\title{Recommendation Functions}
\usage{
kclosestitems(CF, Id_i, k = 10)
}
\arguments{
\item{CF}{A CF objec}

\item{Id_i}{the item Id}

\item{k}{an integer}
}
\description{
Functions that provide items to be recommended to system users.
}
\examples{
objectCF_r <- CFbuilder(Data = movies[1:500,], Datatype = "ratings", 
similarity = "pearson")
kclosestitems(CF = objectCF_r, Id_i = "The Lego Movie")
kclosestitems(CF = objectCF_r, Id_i = "Lincoln", k=5)
}
\author{
Jessica Kubrusly
}
