% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movie ratings by users}
\format{
A data frame with 7276 rows and 3 variables:
\describe{
  \item{Id Users}{Users identifier. Numbers 1 to 526. }
  \item{Id Items}{
  Movies identifier. Movies list:
 \enumerate{
   \item Iron Man 3
   \item Despicable Me 2
   \item My Mom Is a Character
   \item Fast & Furious 6
   \item The Wolverine
   \item Thor: The Dark World
   \item Hansel & Gretel: Witch Hunters
   \item Wreck-It Ralph
   \item Monsters University
   \item The Hangover Part III
   \item Vai Que Dá Certo
   \item Meu Passado me Condena
   \item We’re So Young
   \item Brazilian Western
   \item O Concurso
   \item Mato sem Cachorro
   \item Cine Holliudy
   \item Odeio o Dia dos Namorados
   \item Argo
   \item Django Unchained
   \item Life of Pi
   \item Lincoln
   \item Zero Dark Thirty
   \item Les Miserables
   \item Silver Linings Playbook
   \item Beasts of the Southern Wild
   \item Amour
   \item A Royal Affair
   \item American Hustle
   \item Capitain Phillips
   \item 12 Years a Slave
   \item Dallas Buyers Club
   \item Gravity
   \item Her
   \item Philomena
   \item The Wolf of Wall Street
   \item The Hunt
   \item Frozen
   \item Till Luck Do Us Part 2
   \item Muita Calma Nessa Hora 2
   \item Paranormal Activity: The Marked Ones
   \item I, Frankenstein,
   \item The Legend of Tarzan
   \item The Book Thief
   \item The Lego Movie, , ,
   \item Walking With Dinosaurs
   \item The Hunger Games: Catching Fire
   \item Blue Is The Warmest Color
   \item Reaching for the Moon
   \item The Hobbit: The Desolation of Smaug
   }
  }
  \item{Ratings}{Movie ratings by users. The ratings follows the Likert scale: 1 to 5.}
}
}
\usage{
movies
}
\description{
A dataset containing 7276 ratings for 50 movies by 526 users. This database was created by Giglio (2014).
}
\references{
Giglio , J. C. (2014). Recomendação de Filmes Utilizando Filtragem Colaborativa [Recommending Films Using Collaborative Filtering]. Undergraduate thesis - Universidade Federal Fluminense.
}
\keyword{datasets}
