% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topkitems.R
\name{topkitems}
\alias{topkitems}
\title{Recommendation Function}
\usage{
topkitems(CF, Id_u, k = 10, type = "user")
}
\arguments{
\item{CF}{A CF objec}

\item{Id_u}{the user Id}

\item{k}{an integer}

\item{type}{"user" or "item"}
}
\description{
Functions that provide the top k items
to be recommended to the user Id_u.
}
\examples{
objectCF_r <- CFbuilder(Data = movies[1:500,], Datatype = "ratings", 
similarity = "pearson")
u1 = rownames(objectCF_r$MU)[1]
topkitems(CF=objectCF_r,Id_u = u1)
u2 = rownames(objectCF_r$MU)[2]
topkitems(CF=objectCF_r,Id_u = u2)
}
\author{
Jessica Kubrusly
}
