% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topkusers.R
\name{topkusers}
\alias{topkusers}
\title{Recommendation Function}
\usage{
topkusers(CF, Id_i, k = 10, type = "user")
}
\arguments{
\item{CF}{A CF objec}

\item{Id_i}{the item Id}

\item{k}{an integer}

\item{type}{"user" or "item"}
}
\description{
Functions that provide the top k users
to recommend the item Id_i.
}
\examples{
objectCF_r <- CFbuilder(Data = movies[1:500,], Datatype = "ratings", 
similarity = "pearson")
colnames(objectCF_r$MU)
topkusers(CF = objectCF_r, Id_i = "The Lego Movie")
topkusers(CF = objectCF_r, Id_i = "Her")
}
\author{
Jessica Kubrusly
}
